backend/requirements.txt
Django==5.2.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.0
django-filter==25.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
Pillow==11.3.0
